Split Python Types into their own c module like they should be.
